負にならない(なるべきでない)変数、関数を unsigned int に変更
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@6982 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ ttssh2/ttxssh/pkt.c
@@ -174,7 +174,7 @@ int PKT_recv(PTInstVar pvar, char *buf, int buflen)
 			pvar->pkt_state.datastart += i;
 			pvar->pkt_state.datalen -= i;
 		}
-		else if (pvar->pkt_state.seen_server_ID && pvar->pkt_state.datalen >= (unsigned int) SSH_get_min_packet_size(pvar)) {
+		else if (pvar->pkt_state.seen_server_ID && pvar->pkt_state.datalen >= SSH_get_min_packet_size(pvar)) {
 			char *data = pvar->pkt_state.buf + pvar->pkt_state.datastart;
 			uint32 padding;
 			uint32 pktsize;
